Only allow specific applications to run in headless mode
We're not going to support the headless mode anymore. We need to keep the number of web
process running in the whole system at the same time on a minimum and spawning more than
one per application doesn't make sense. If the application needs to do some background
activity it should use a service for this and trigger it through the activitymanager. To
respect already existing legacy applications we add a static list of applications with
this change we will grant the headless mode. The list should stay static and there should
be no need to extend it.

Signed-off-by: Simon Busch <morphis@gravedo.de>